requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
coreapi==2.3.0
coreschema==0.0.4
cryptography==1.8.1
Django==1.10.6
django-cors-headers==2.0.2
django-debug-toolbar==1.7
django-extensions==1.7.8
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
Fabric==1.13.1
Fabric3==1.13.1.post1
idna==2.5
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.4.2
PyMySQL==0.7.10
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
ujson==1.35
uritemplate==3.0.0
uWSGI==2.0.15
